Prevent unsafe narrowing: misc. dependencies of ui/views
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3712273
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017870}

+++ cc/paint/display_item_list.h
@@ -175,7 +175,9 @@ class CC_PAINT_EXPORT DisplayItemList
   void EmitTraceSnapshot() const;
   void GenerateDiscardableImagesMetadata();
 
-  gfx::Rect VisualRectForTesting(int index) { return visual_rects_[index]; }
+  gfx::Rect VisualRectForTesting(int index) {
+    return visual_rects_[static_cast<size_t>(index)];
+  }
 
   // Generate a PaintRecord from this DisplayItemList, leaving |this| in
   // an empty state.
